requirements.txt
discord.py
openai>=1.1.1
aiohttp